Fix providers
