+++ lib/page/game/component/side_pot.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:poker_chip/model/entity/side_pot/side_pot_entity.dart';
-import 'package:poker_chip/provider/presentation_providers.dart';
+import 'package:poker_chip/provider/presentation/pot.dart';
 
 class SidePotsWidget extends ConsumerWidget {
   const SidePotsWidget(this.isHost, {super.key});
