+++ lib/page/game/participant/component/participant_action_button.dart
@@ -4,6 +4,7 @@ import 'package:poker_chip/model/entity/action/action_entity.dart';
 import 'package:poker_chip/model/entity/message/message_entity.dart';
 import 'package:poker_chip/page/game/host/host_page.dart';
 import 'package:poker_chip/provider/presentation/opt_id.dart';
+import 'package:poker_chip/provider/presentation/peer.dart';
 import 'package:poker_chip/provider/presentation/player.dart';
 import 'package:poker_chip/provider/presentation_providers.dart';
 import 'package:poker_chip/util/enum/action.dart';
